CN0429: Add support for CN0429 shield
The CN0429_example is a demo using the EVAL-CN0428-EBZ and the
EVAL-M355-ARDZ-INT boards as a solution for detecting toxic gases. By
utilizing built-in diagnostics features (such as impedance spectroscopy or
bias voltage pulsing and ramping) it is possible to inspect sensor health,
compensate for accuracy drift due to aging or temperature, and estimate the
remaining lifetime of the sensor.

Signed-off-by: Mircea Caprioru <mircea.caprioru@analog.com>